fix errors in struct field types

+++ pkg/formats/test-formats.go
@@ -78,96 +78,96 @@ type TestFormat12Bad struct {
 type TestFormat13Bad struct {
 	TestWord0  `word:"24"`
 	TestWord1  `word:"32"`
-	TestField2 byte `bitfield:"8,0"` // bitfield in format
+	TestField2 uint `bitfield:"8,0"` // bitfield in format
 }
 
 type TestWord0 struct {
 	//                                   2   1         0
 	//                                   321098765432109876543210
-	TestField0 byte `bitfield:"8,16"` // |------|
-	TestField1 byte `bitfield:"8,8"`  //         |------|
-	TestField2 byte `bitfield:"8,0"`  //                 |------|
+	TestField0 uint `bitfield:"8,16"` // |------|
+	TestField1 uint `bitfield:"8,8"`  //         |------|
+	TestField2 uint `bitfield:"8,0"`  //                 |------|
 }
 
 type TestWord1 struct {
 	//                                   3 2         1         0
 	//                                   10987654321098765432109876543210
-	TestField0 byte `bitfield:"5,27"` // |---|
-	TestField1 byte `bitfield:"7,20"` //      |-----|
-	TestField2 byte `bitfield:"9,11"` //             |-------|
-	TestField3 byte `bitfield:"11,0"` //                      |---------|
+	TestField0 uint `bitfield:"5,27"` // |---|
+	TestField1 uint `bitfield:"7,20"` //      |-----|
+	TestField2 uint `bitfield:"9,11"` //             |-------|
+	TestField3 uint `bitfield:"11,0"` //                      |---------|
 }
 
 type TestWord2 struct {
 	//                                    3         2         1         0
 	//                                    9876543210987654321098765432109876543210
-	TestField0 byte `bitfield:"3,37"`  // |-|
-	TestField1 byte `bitfield:"10,27"` //    |--------|
-	TestField2 byte `bitfield:"6,21"`  //              |----|
-	TestField3 byte `bitfield:"13,8"`  //                    |-----------|
-	TestField4 byte `bitfield:"8,0"`   //                                 |------|
+	TestField0 uint `bitfield:"3,37"`  // |-|
+	TestField1 uint `bitfield:"10,27"` //    |--------|
+	TestField2 uint `bitfield:"6,21"`  //              |----|
+	TestField3 uint `bitfield:"13,8"`  //                    |-----------|
+	TestField4 uint `bitfield:"8,0"`   //                                 |------|
 }
 
 type TestWord3Bad struct {
 	// gap between fields                2   1         0
 	//                                   321098765432109876543210
-	TestField0 byte `bitfield:"6,18"` // |----|
-	TestField1 byte `bitfield:"8,8"`  //         |------|
-	TestField2 byte `bitfield:"8,0"`  //                 |------|
+	TestField0 uint `bitfield:"6,18"` // |----|
+	TestField1 uint `bitfield:"8,8"`  //         |------|
+	TestField2 uint `bitfield:"8,0"`  //                 |------|
 }
 
 type TestWord4Bad struct {
 	// overlapping fields                3 2         1         0
 	//                                   10987654321098765432109876543210
-	TestField0 byte `bitfield:"5,27"` // |---|
-	TestField1 byte `bitfield:"9,18"` //      |-------|
-	TestField2 byte `bitfield:"9,11"` //             |-------|
-	TestField3 byte `bitfield:"11,0"` //                      |---------|
+	TestField0 uint `bitfield:"5,27"` // |---|
+	TestField1 uint `bitfield:"9,18"` //      |-------|
+	TestField2 uint `bitfield:"9,11"` //             |-------|
+	TestField3 uint `bitfield:"11,0"` //                      |---------|
 }
 
 type TestWord5Bad struct {
 	// gap and overlap between fields     3         2         1         0
 	//                                    9876543210987654321098765432109876543210
-	TestField0 byte `bitfield:"3,29"`  // |-|
-	TestField1 byte `bitfield:"10,19"` //    |--------|
-	TestField2 byte `bitfield:"6,11"`  //                |----|
-	TestField3 byte `bitfield:"13,0"`  //                    |-----------|
-	TestField4 byte `bitfield:"8,0"`   //                                 |------|
+	TestField0 uint `bitfield:"3,29"`  // |-|
+	TestField1 uint `bitfield:"10,19"` //    |--------|
+	TestField2 uint `bitfield:"6,11"`  //                |----|
+	TestField3 uint `bitfield:"13,0"`  //                    |-----------|
+	TestField4 uint `bitfield:"8,0"`   //                                 |------|
 }
 
 type TestWord6Bad struct {
 	// missing field                    2   1         0
 	//                                  321098765432109876543210
-	TestField1 byte `bitfield:"8,8"` //         |------|
-	TestField2 byte `bitfield:"8,0"` //                 |------|
+	TestField1 uint `bitfield:"8,8"` //         |------|
+	TestField2 uint `bitfield:"8,0"` //                 |------|
 }
 
 type TestWord7Bad struct {
 	// missing struct tag                3 2         1         0
 	//                                   10987654321098765432109876543210
-	TestField0 byte `bitfield:"5,27"` // |---|
-	TestField1 byte
-	TestField2 byte `bitfield:"9,11"` //             |-------|
-	TestField3 byte `bitfield:"11,0"` //                      |---------|
+	TestField0 uint `bitfield:"5,27"` // |---|
+	TestField1 uint
+	TestField2 uint `bitfield:"9,11"` //             |-------|
+	TestField3 uint `bitfield:"11,0"` //                      |---------|
 }
 
 type TestWord8Bad struct {
 	// repeated fields                    3         2         1         0
 	//                                    9876543210987654321098765432109876543210
-	TestField0 byte `bitfield:"3,37"`  // |-|
-	TestField1 byte `bitfield:"10,27"` //    |--------|
-	TestField2 byte `bitfield:"6,21"`  //              |----|
-	TestField3 byte `bitfield:"6,21"`  //              |----|
-	TestField4 byte `bitfield:"13,8"`  //                    |-----------|
-	TestField5 byte `bitfield:"8,0"`   //                                 |------|
+	TestField0 uint `bitfield:"3,37"`  // |-|
+	TestField1 uint `bitfield:"10,27"` //    |--------|
+	TestField2 uint `bitfield:"6,21"`  //              |----|
+	TestField3 uint `bitfield:"6,21"`  //              |----|
+	TestField4 uint `bitfield:"13,8"`  //                    |-----------|
+	TestField5 uint `bitfield:"8,0"`   //                                 |------|
 }
 
 type TestWord9Bad struct {
 	// total length not multiple of 8    1         0
 	//                                   98765432109876543210
-	TestField0 byte `bitfield:"8,12"` // |------|
-	TestField1 byte `bitfield:"8,4"`  //         |------|
-	TestField2 byte `bitfield:"4,0"`  //                 |--|
+	TestField0 uint `bitfield:"8,12"` // |------|
+	TestField1 uint `bitfield:"8,4"`  //         |------|
+	TestField2 uint `bitfield:"4,0"`  //                 |--|
 }
 
 type TestWord10Bad struct {
@@ -175,13 +175,13 @@ type TestWord10Bad struct {
 
 type TestWord11Bad struct {
 	TestField0 string `bitfield:"8,16"` // unsupported field type
-	TestField1 byte   `bitfield:"8,8"`
-	TestField2 byte   `bitfield:"8,0"`
+	TestField1 uint   `bitfield:"8,8"`
+	TestField2 uint   `bitfield:"8,0"`
 }
 
 type TestWord12Bad struct {
-	TestField0 byte `bitfield:"5,27"`
-	TestField1 byte `bitfield:"7,20"`
+	TestField0 uint `bitfield:"5,27"`
+	TestField1 uint `bitfield:"7,20"`
 	TestField2 byte `bitfield:"9,11"` // field size overflows type
 	TestField3 uint `bitfield:"11,0"`
 }
